server/requirements.txt
captcha==0.2.1
Django==1.10.5
djangorestframework==3.5.3

